☘️ refactor [#11.5.7]: 4차 개선 - 아키텍처 계층 정립 및 URL/렌더링 버그 수정
- **[types]**: shared chat 타입 레이어 신설 및 SourceItem 이동 (계층 구조 위반 해결).
- **[i18n-utils]**: createPathWithLocale에서 Query/Hash 소실 방지 로직 도입.
- **[MessageBubble]**: props 비교 시 role 체크 추가하여 렌더링 정합성 확보.
- **[SourcePanel/chat-utils]**: 신규 타입 레이어 참조로 변경.

🔗 Related:
- Issue [#775]
- ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/801#pullrequestreview-3973271464)
