+++ web_ui/src/lib/i18n-utils.ts
@@ -9,17 +9,20 @@ import { type Locale, locales } from '@/i18n/config';
 export function createPathWithLocale(pathname: string, newLocale: Locale): string {
   if (!pathname) return '/';
 
-  const segments = pathname.split('/');
+  // [PR 리뷰 반영] pathname에 쿼리 스트링이나 해시가 포함된 경우를 분리 처리
+  const [purePathname, ...rest] = pathname.split(/(\?|#)/);
+  const suffix = rest.join('');
+
+  const segments = purePathname.split('/');
   
   // Helper Check: 루트 경로인지 확인
-  // pathname이 '/' 일 때 split('/') 결과는 ['', ''] 입니다.
   const isRootPath = segments.length === 2 && segments[1] === '';
 
   if (isRootPath) {
     // 루트 경로인 경우, 빈 문자열 세그먼트를 새 로케일로 교체하여
     // 불필요한 Trailing Slash 방지 (예: /ko/ 대신 /ko 생성)
     segments[1] = newLocale;
-    return segments.join('/');
+    return segments.join('/') + suffix;
   }
 
   // 동적으로 현재 로케일 세그먼트 위치 탐색
@@ -34,5 +37,5 @@ export function createPathWithLocale(pathname: string, newLocale: Locale): strin
     segments.splice(1, 0, newLocale);
   }
 
-  return segments.join('/');
+  return segments.join('/') + suffix;
 }
